Fix loading SQS queues when using binary state format
Some of the queue attributes are modelled as python lambdas, so they are pickled weirdly by `dill`. In localstack 4.9 (possibly because of the underlying python update), this actually causes a segfault when loading state from a previous version of localstack-persist!

To work-around that, always replace any callable (i.e. lambda) attributes of the SQS queue that were loaded from persisted state.